requirements.txt
--index-url https://download.pytorch.org/whl/cu118
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.20.0
opencv-python>=4.5.3
matplotlib>=3.4.3
scikit-image>=0.18.3
tqdm>=4.62.3
pyyaml>=6.0
tensorboard>=2.7.0
einops>=0.4.1
timm>=0.5.4
lpips>=0.1.4
flask>=2.0.0
werkzeug>=2.0.0
streamlit>=1.20.0
gunicorn>=20.1.0
